requirements.txt
aiogram
sqlalchemy
aiohttp
python-dotenv
asyncpg

